+++ src/main/java/com/hhp/ecommerce/presentation/facade/BalanceFacadeImpl.java
@@ -1,4 +1,4 @@
-package com.hhp.ecommerce.presentation;
+package com.hhp.ecommerce.presentation.facade;
 
 import org.springframework.stereotype.Service;
